Make static functions non-member functions
