+++ test/include/psen_scan_v2/test_data_helper.h
@@ -13,6 +13,9 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+#ifndef PSEN_SCAN_V2_TEST_TEST_DATA_HELPER_H
+#define PSEN_SCAN_V2_TEST_TEST_DATA_HELPER_H
+
 #include <algorithm>
 #include <cstdint>
 #include <memory>
@@ -24,43 +27,26 @@
 #include "psen_scan_v2_standalone/util/timestamp.h"
 
 #include "psen_scan_v2/test_data.h"
-#include "psen_scan_v2/test_data_assembler.h"
-#include "psen_scan_v2/udp_data_reader.h"
+#include "psen_scan_v2/udp_data_helper.h"
 
 namespace psen_scan_v2_test
+{
+using namespace psen_scan_v2_standalone;
+
+namespace test_data
 {
 static constexpr uint16_t FIRST_FROM_THETA{ 2500 };  // change to 1 after scan counter fix in firmware
 static constexpr uint16_t LAST_FROM_THETA{ 2000 };   // change to 2500 after scan counter fix in firmware
 
-int64_t secToNSec(const double& sec)
+static int64_t secToNSec(const double& sec)
 {
   double nsec{ sec * 1000000000.0 };
   return static_cast<int64_t>(nsec >= 0 ? nsec + 0.5 : nsec - 0.5);
 }
 
-std::unique_ptr<TestData> TestDataAssembler::assemble(const ScannerConfiguration& scanner_config,
-                                                      const int64_t scanner_run_duration_sec,
-                                                      const std::string& udp_data_filename,
-                                                      const uint16_t udp_port)
-{
-  std::unique_ptr<TestData> test_data_ptr{ new TestData() };
-  runScanner(scanner_config, scanner_run_duration_sec, [&test_data_ptr](const LaserScan& scan) {
-    extractDataFromScan(*test_data_ptr, scan);
-  });
-  PSENSCAN_INFO("TestDataAssembler", "Extracted data from {} laserscan callbacks.", test_data_ptr->size());
-
-  UdpData udp_data;
-  UdpDataReader::read(udp_data_filename, udp_port, udp_data);
-  addUdpData(*test_data_ptr, udp_data);
-
-  removeIncompleteData(*test_data_ptr);
-  sortWithRespectToScanCounter(*test_data_ptr);
-  return test_data_ptr;
-}
-
-void TestDataAssembler::runScanner(const ScannerConfiguration& scanner_config,
-                                   const int64_t scanner_run_duration_sec,
-                                   const IScanner::LaserScanCallback& laserscan_callback)
+static void runScanner(const ScannerConfiguration& scanner_config,
+                       const int64_t scanner_run_duration_sec,
+                       const IScanner::LaserScanCallback& laserscan_callback)
 {
   ScannerV2 scanner(scanner_config, laserscan_callback);
   scanner.start();
@@ -68,7 +54,7 @@ void TestDataAssembler::runScanner(const ScannerConfiguration& scanner_config,
   scanner.stop();
 }
 
-void TestDataAssembler::addUdpData(TestData& test_data, const UdpData& udp_data)
+static void addUdpData(TestData& test_data, const udp_data::UdpData& udp_data)
 {
   for (const auto& udp_datum : udp_data)
   {
@@ -89,23 +75,46 @@ void TestDataAssembler::addUdpData(TestData& test_data, const UdpData& udp_data)
   }
 }
 
-void TestDataAssembler::removeIncompleteData(TestData& test_data)
+static void removeIncompleteData(TestData& test_data)
 {
   test_data.erase(
       std::remove_if(test_data.begin(), test_data.end(), [](const auto& datum) { return !datum.isComplete(); }),
       test_data.end());
 }
 
-void TestDataAssembler::sortWithRespectToScanCounter(TestData& test_data)
+static void sortWithRespectToScanCounter(TestData& test_data)
 {
   std::sort(test_data.begin(), test_data.end(), [](const auto& datum1, const auto& datum2) {
     return datum1.scanCounter() < datum2.scanCounter();
   });
 }
 
-void TestDataAssembler::extractDataFromScan(TestData& test_data, const LaserScan& scan)
+static void extractDataFromScan(TestData& test_data, const LaserScan& scan)
 {
   test_data.push_back(TestDatum(scan.getScanCounter(), scan.getTimestamp(), util::getCurrentTime()));
 }
 
+static std::unique_ptr<TestData> assemble(const ScannerConfiguration& scanner_config,
+                                          const int64_t scanner_run_duration_sec,
+                                          const std::string& udp_data_filename,
+                                          const uint16_t udp_port)
+{
+  std::unique_ptr<TestData> test_data_ptr{ new TestData() };
+  runScanner(scanner_config, scanner_run_duration_sec, [&test_data_ptr](const LaserScan& scan) {
+    extractDataFromScan(*test_data_ptr, scan);
+  });
+  PSENSCAN_INFO("TestDataAssembler", "Extracted data from {} laserscan callbacks.", test_data_ptr->size());
+
+  udp_data::UdpData udp_data;
+  udp_data::read(udp_data_filename, udp_port, udp_data);
+  addUdpData(*test_data_ptr, udp_data);
+
+  removeIncompleteData(*test_data_ptr);
+  sortWithRespectToScanCounter(*test_data_ptr);
+  return test_data_ptr;
+}
+
+}  // namespace test_data
 }  // namespace psen_scan_v2_test
+
+#endif  // PSEN_SCAN_V2_TEST_TEST_DATA_HELPER_H
